refactor: centralize blob download helper
