+++ web/src/lib/warp/useWarpTransfer.ts
@@ -27,6 +27,7 @@
  */
 
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { saveBlob } from "../saveBlob";
 import { SignalingClient, type SignalData } from "./signaling";
 import { deviceName } from "./deviceName";
 import {
@@ -267,19 +268,6 @@ async function uniqueName(used: Set<string>, name: string, dir?: FsDirHandle): P
     }
     candidate = `${stem} (${n})${ext}`;
   }
-}
-
-/** Trigger a browser download of a blob via a transient object-URL anchor. */
-function saveBlob(blob: Blob, filename: string): void {
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  document.body.appendChild(a);
-  a.click();
-  a.remove();
-  // Revoke on the next tick so the click has time to start the download.
-  setTimeout(() => URL.revokeObjectURL(url), 1000);
 }
 
 export function useWarpTransfer(joinCode?: string): UseWarpTransfer {
